added downloadFile util
